Revert "fixing ObservableError and ObserveErrorListener ambiguity"
This reverts commit 5e40a88428b82f47972bab67e74f941c407a185f.
