+++ packages/functions/src/functions.ts
@@ -4,7 +4,6 @@ import { BasedFunctionClient } from './client.js'
 import { BasedDataStream } from './stream.js'
 import { Authorize } from './auth.js'
 import type { Plugin } from 'esbuild'
-import { BasedErrorCode, BasedErrorData } from '@based/errors'
 
 export type ObservableUpdateFunction<K = any> = (
   data: K,
@@ -16,12 +15,7 @@ export type ObservableUpdateFunction<K = any> = (
   isDeflate?: boolean,
 ) => void
 
-export type ObservableError =
-  | BasedErrorData<BasedErrorCode.FunctionError>
-  | BasedErrorData<BasedErrorCode.FunctionIsWrongType>
-  | BasedErrorData<BasedErrorCode.FunctionNotFound>
-
-export type ObserveErrorListener = (err: ObservableError) => void
+export type ObserveErrorListener = (err: any) => void
 
 export type HttpHeaders = {
   [header: string]: number | string | (string | number)[]
